Unify RowWritable and RecordWrapperWritable
